# api-docs/swagger.yaml
openapi: 3.0.0
info:
  title: CLM-Core
  version: 1.0.0
  description: API endpoints the clm-core module offers
servers:
  - url: '{scheme}://{hostname}:{port}{path}'
    description: The production API server
    variables:
      hostname:
        default: localhost
      port:
        default: '5008'
      path:
        default: ''
      scheme:
        default: http
security:
  - bearerAuth: []
paths:
...
  '/core/mgmt/groups/{groupId}/users/{userId}':
    post:
      tags:
        - mgmt-groups
        - mgmt
      description: >-
        Add user can add a target user to a group if the user is admin and the
        target user in a child-group of the admin group
      summary: 'Add user to a group [Minimum Role: ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
        - in: path
          name: groupId
          description: The id of the group
          required: true
          schema:
            type: string
        - in: path
          name: userId
          description: The userId of the group
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully added user to group
    delete:
      tags:
        - mgmt-groups
        - mgmt
      description: >-
        A user can delete a target user from a group if the user is admin and
        the target user in a child-group of the admin group
      summary: 'Remove user from a group [Minimum Role: ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
        - in: path
          name: groupId
          description: The id of the group
          required: true
          schema:
            type: string
        - in: path
          name: userId
          description: The userId of the group
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successfully removed user from group
  /core/mgmt/groups:
    get:
      tags:
        - mgmt-groups
        - mgmt
      description: A user can access all his groups and subgroups
      summary: 'Get all groups [Minimum Role: LEARNER]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: Successfully removed user from group
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/group'
    post:
      tags:
        - mgmt-groups
        - mgmt
      description: A user can create a group if hes atleast ADMIN in another group
      summary: 'Create group [Minimum Role: ADMIN]'
      requestBody:
        content:
          application/json:
            schema:
              properties:
                displayName:
                  type: string
                  description: how the group should be named
                  default: Example-Group
                role:
                  type: string
                  enum:
                    - LEARNER
                    - INSTRUCTOR
                    - ADMIN
                  description: The role you want to give the group
                  default: LEARNER
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: Successfully removed user from group
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/group'
  /core/mgmt/groups/relations:
    get:
      tags:
        - mgmt-groups
      description: Getting all relations of the groups the user is part of
      summary: 'Get group-relations [Minimum Role: LEARNER]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: Successfully added group to a group
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/relation'
  '/core/mgmt/groups/{groupId}':
    patch:
      tags:
        - mgmt-groups
        - mgmt
      description: A user can update a group if he has admin privileges on that group
      summary: 'Updating a specific group [Minimum Role: ADMIN]'
      requestBody:
        content:
          application/json:
            schema:
              properties:
                displayName:
                  type: string
                  description: how the group should be named
                  default: Example-Group
                role:
                  type: string
                  enum:
                    - LEARNER
                    - INSTRUCTOR
                    - ADMIN
                  description: The role you want to give the group
                  default: LEARNER
      parameters:
        - $ref: '#/components/parameters/accessToken'
        - in: path
          required: true
          name: groupId
          description: the id of the group
          schema:
            type: string
      responses:
        '200':
          description: Successfully updated a group
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/group'
    delete:
      tags:
        - mgmt-groups
        - mgmt
      description: A user can delete a group if he has admin privileges on that group
      summary: 'Delete a specific group [Minimum Role: ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
        - in: path
          required: true
          name: groupId
          description: the id of the group
          schema:
            type: string
      responses:
        '204':
          description: Successfully deleted a group
  '/core/mgmt/groups/{groupId}/groups/{childGroupId}':
    post:
      tags:
        - mgmt-groups
        - mgmt
      description: >-
        A user can add a group to a group if he has has admin privileges on the
        from group and the target group
      summary: 'Add group to group [Minimum Role: ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
        - in: path
          name: groupId
          description: the id of the group
          required: true
          schema:
            type: string
        - in: path
          name: childGroupId
          description: the id of the child-group
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successfully added a group to a group
...
  /core/users/me/groups:
    get:
      tags:
        - pblc
      description: Get the groups the user is part of
      summary: Get the groups the user is part of
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: User groups
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - properties:
                        role:
                          type: string
                          enum:
                            - LEARNER
                            - INSTRUCTOR
                            - ADMIN
                            - SUPER-ADMIN
                          description: the role the user has on this group
                    - $ref: '#/components/schemas/group'
  '/core/users/{userId}/groups':
    get:
      tags:
        - pblc
      description: Get the groups the user is part of
      summary: Get the groups the user is part of
      parameters:
        - $ref: '#/components/parameters/accessToken'
        - in: path
          name: userId
          description: id of the user
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User groups
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - properties:
                        role:
                          type: string
                          enum:
                            - LEARNER
                            - INSTRUCTOR
                            - ADMIN
                            - SUPER-ADMIN
                          description: the role the user has on this group
                    - $ref: '#/components/schemas/group'
...
components:
  schemas:
    relation:
      type: object
      properties:
        fromType:
          type: string
          description: The type of the node
          default: fromTypeNode
        toType:
          type: string
          description: The type of the target node
          default: toTypeNode
        fromId:
          type: string
          description: The id of the node
          default: fromNodeId
        toId:
          type: string
          description: The id of the target node
          default: toNodeId
        order:
          type: number
          description: >-
            The order of the relation. Used for example ordering the enrollments
            of a group/user
          default: 0
...
    group:
      type: object
      properties:
        _id:
          type: string
          description: id of the group
          default: jaiosdjfiojifj2341dij
        displayName:
          type: string
          description: how the group should be displayed
          default: Example-Group
        updatedAt:
          type: string
          description: when the group was last updated
          default: 10-10-2020
        createdAt:
          type: string
          description: id of the group
          default: 10-10-2022
...
  parameters:
    accessToken:
      name: x-access-token
      in: header
      description: The access token
      required: true
      example: exampleAccessToken
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    refreshAuth:
      type: apiKey
      in: header
      name: x-refresh-token
...